Add Time proof Non-Fungible Token Standard
## Simple Summary
This is a standard for time-stamp provable NFTs, which extends [ERC-721]([./eip-721.md](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-721.md)).
Enables NFT to mapping and anchor things that have time features. 
And also the time-stamp proof of NFTs can be the condition of interaction with smart contracts.
It is called Time proof Non-fungible Token and is referred to as tpNFT in the subsequent descriptions.
## Abstract
Based on the tokenURI parameter of ERC721, the plain text parameter and the encryption parameter are added, and the time proof is set for the update of these three parameters.
With these improvements, we can provide new possibilities for NFT, enabling NFT to map and capture time-level value.
Here are some brand new applications that can be achieved by tpNFT based on this extension. I believe you can find more and more interesting applications to bring more functions to the blockchain.

- Governance: As a proof of timestamp for governance voting, it is used to verify the validity of the vote and obtain governance incentives.
- Invention and creation: The idea is stored on the chain to prove that the idea existed before a certain point in time. And you can temporarily hide the specific technical implementation details through Hash encryption. Open source projects： Similarly to previous one to used to prove the originality of their products and leave a proof of their own open source.
- Scientific discoveries: Publish the publicly available content of relevant scientific discoveries in plain text, and fix some temporarily undisclosed demonstration details through encrypted parameters. Engrave the brilliance of human wisdom on the blockchain. The timestamp proves that it will not lie.
- Prophecy: Make predictions and cast them into tpNFT with time proof ability, as proof of personal ability, or as a certificate for receiving bonuses.
- Identity: Participating in a series of on-chain behaviors through tpNFT, including the prophecy mentioned in the previous point, revolves around tpNFT's timestamp proof, which has become a natural proof of identity.
- Legacy: Binding assets to a tpNFT, others can receive part or all of the assets by submitting the preimage of encryption parameter(Hash or HashTree). We have another EIP to solve the front-running attacks.
- Dynamic NFT: With tpNFT, we only need a little improvement than can get a brand new type of NFT which state on the chain will automatically change over time without consuming any gas fees. Such features can open up new areas for games, installation art, financial contracts, and so on.
   
Summarize: From a philosophical point of view, The NFT of the ERC721 standard is mainly used to anchor and map things related to space, while tpNFT is used for related to time.